Real estate project1
This project we use real estate dataset to estimate the results by using OLS and ML methods, but we do one more things is that we compare the results by preparing the data  or not before running the OLS and ML regression. So that's why we split the process as 2 parts.